Rename OutboxService to ProtobufOutboxService
