+++ outbox-kafka-spring-reactive/src/test/java/one/tomorrow/transactionaloutbox/reactive/AbstractIntegrationTest.java
@@ -18,7 +18,7 @@ package one.tomorrow.transactionaloutbox.reactive;
 import one.tomorrow.transactionaloutbox.reactive.model.OutboxLock;
 import one.tomorrow.transactionaloutbox.reactive.repository.OutboxLockRepository;
 import one.tomorrow.transactionaloutbox.reactive.service.OutboxLockService;
-import one.tomorrow.transactionaloutbox.reactive.service.OutboxService;
+import one.tomorrow.transactionaloutbox.reactive.service.ProtobufOutboxService;
 import org.flywaydb.test.junit5.annotation.FlywayTestExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.slf4j.Logger;
@@ -36,7 +36,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
         OutboxLockRepository.class,
         OutboxLock.class,
         OutboxLockService.class,
-        OutboxService.class,
+        ProtobufOutboxService.class,
         IntegrationTestConfig.class
 })
 @Testcontainers
